Salman

+++ Eatr/app/src/main/java/sk/greate43/eatr/adaptors/SellFoodRecyclerViewAdaptor.java
@@ -9,6 +9,8 @@ import android.view.ViewGroup;
 
 import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseUser;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.storage.FirebaseStorage;
 import com.google.firebase.storage.StorageReference;
@@ -17,9 +19,10 @@ import java.util.ArrayList;
 
 import sk.greate43.eatr.R;
 import sk.greate43.eatr.activities.SellerActivity;
-import sk.greate43.eatr.entities.Seller;
+import sk.greate43.eatr.entities.Food;
 import sk.greate43.eatr.holders.SellFoodRecyclerViewHolder;
 import sk.greate43.eatr.interfaces.SwipeListener;
+import sk.greate43.eatr.utils.Constants;
 
 /**
  * Created by great on 11/12/2017.
@@ -28,22 +31,28 @@ import sk.greate43.eatr.interfaces.SwipeListener;
 public class SellFoodRecyclerViewAdaptor extends RecyclerView.Adapter<SellFoodRecyclerViewHolder> implements SwipeListener {
 
     private static final String TAG = "SellFoodRecyclerViewAda";
+
+
     DatabaseReference mDatabaseReference;
     private StorageReference storageReference;
-    private ArrayList<Seller> sellers;
+    private FirebaseAuth mAuth;
+    private FirebaseUser user;
+    private ArrayList<Food> foods;
     private LayoutInflater inflater;
     private SellerActivity sellerActivity;
 
     public SellFoodRecyclerViewAdaptor(SellerActivity sellerActivity, DatabaseReference mDatabaseReference) {
         this.sellerActivity = sellerActivity;
         inflater = sellerActivity.getLayoutInflater();
-        sellers = new ArrayList<>();
+        foods = new ArrayList<>();
         this.mDatabaseReference = mDatabaseReference;
+        mAuth=FirebaseAuth.getInstance();
+        user=mAuth.getCurrentUser();
 
     }
 
-    public ArrayList<Seller> getSellers() {
-        return sellers;
+    public ArrayList<Food> getFoods() {
+        return foods;
     }
 
     @Override
@@ -56,25 +65,26 @@ public class SellFoodRecyclerViewAdaptor extends RecyclerView.Adapter<SellFoodRe
 
     @Override
     public void onBindViewHolder(SellFoodRecyclerViewHolder holder, int position) {
-        if (sellers == null || sellers.size() == 0) {
+        if (foods == null || foods.size() == 0) {
             holder.imgFoodItem.setImageResource(R.drawable.ic_launcher_background);
 
         } else {
-            holder.populate(sellerActivity, sellers.get(position));
+
+            holder.populate(sellerActivity, foods.get(position));
         }
     }
 
     @Override
     public int getItemCount() {
-        return ((sellers != null && sellers.size() != 0) ? sellers.size() : 1);
+        return ((foods != null && foods.size() != 0) ? foods.size() : 1);
     }
 
 
     public void clear() {
-        int size = sellers.size();
+        int size = foods.size();
         if (size > 0) {
             for (int i = 0; i < size; i++) {
-                sellers.remove(0);
+                foods.remove(0);
             }
 
             notifyItemRangeRemoved(0, size);
@@ -82,31 +92,39 @@ public class SellFoodRecyclerViewAdaptor extends RecyclerView.Adapter<SellFoodRe
     }
 
     @Override
+    public int getItemViewType(int position) {
+        return super.getItemViewType(position);
+    }
+
+    @Override
     public void onSwipe(int position) {
-        Log.d(TAG, "onSwipe: " + position);
-        Log.d(TAG, "onSwipe: " + sellers.get(position).getDishName());
-        mDatabaseReference.child("eatr").child("greate43").child(sellers.get(position).getDishName()).removeValue();
-        storageReference = FirebaseStorage.getInstance().getReferenceFromUrl(sellers.get(position).getImageUri());
+        if (!foods.isEmpty()) {
+            Log.d(TAG, "onSwipe: " + position);
+            Log.d(TAG, "onSwipe: " + foods.get(position).getDishName());
+            mDatabaseReference.child(Constants.FOOD).child(user.getUid()).child(foods.get(position).getDishName()).removeValue();
+            storageReference = FirebaseStorage.getInstance().getReferenceFromUrl(foods.get(position).getImageUri());
 
-        storageReference.delete().addOnSuccessListener(new OnSuccessListener<Void>() {
-            @Override
-            public void onSuccess(Void aVoid) {
-                // File deleted successfully
-                Log.d(TAG, "onSuccess: deleted file");
-            }
-        }).addOnFailureListener(new OnFailureListener() {
-            @Override
-            public void onFailure(@NonNull Exception exception) {
-                // Uh-oh, an error occurred!
-                Log.d(TAG, "onFailure: did not delete file");
-            }
-        });
+            storageReference.delete().addOnSuccessListener(new OnSuccessListener<Void>() {
+                @Override
+                public void onSuccess(Void aVoid) {
+                    // File deleted successfully
+                    Log.d(TAG, "onSuccess: deleted file");
+                }
+            }).addOnFailureListener(new OnFailureListener() {
+                @Override
+                public void onFailure(@NonNull Exception exception) {
+                    // Uh-oh, an error occurred!
+                    Log.d(TAG, "onFailure: did not delete file");
+                }
+            });
+            removeItem(position);
 
-        removeItem(position);
+        }
+
     }
 
     private void removeItem(int position) {
-        sellers.remove(position);
+        foods.remove(position);
         // notify the item removed by position
         // to perform recycler view delete animations
         // NOTE: don't call notifyDataSetChanged()
